feat(nvim): migrate to lazy.nvim from packer
