--- .config/nvim/after/plugin/alpha.lua
@@ -1,1 +1,0 @@
-require('alpha').setup(require('senyc.launch'))
