--- .config/nvim/after/plugin/glow.lua
@@ -1,8 +1,0 @@
-require('glow').setup {
-  style = 'dark',
-  pager = false, -- This is broken
-  border = 'rounded',
-  width_ratio = 0.8,
-  height_ratio = 0.8,
-  width = 210,
-}
