--- .config/nvim/after/plugin/illuminate.lua
@@ -1,27 +1,0 @@
-require('illuminate').configure({
-  -- providers: provider used to get references in the buffer, ordered by priority
-  providers = {
-    'lsp',
-    'treesitter',
-  },
-  delay = 100,
-  filetype_overrides = {},
-  -- filetypes_denylist: filetypes to not illuminate, this overrides filetypes_allowlist
-  filetypes_denylist = {
-    'NvimTree',
-    'packer',
-    'toggleterm',
-    'TelescopePrompt',
-    'fugitive',
-  },
-  filetypes_allowlist = {},
-  modes_denylist = {},
-  modes_allowlist = {},
-  providers_regex_syntax_denylist = {},
-  providers_regex_syntax_allowlist = {},
-  -- under_cursor: whether or not to illuminate under the cursor
-  under_cursor = true,
-  large_file_cutoff = nil,
-  large_file_overrides = nil,
-  min_count_to_highlight = 1,
-})
