--- .config/nvim/after/plugin/lsp/mason.lua
@@ -1,5 +1,0 @@
-require('mason').setup {
-  ui = {
-    border = 'rounded'
-  }
-}
