--- .config/nvim/after/plugin/undotree.lua
@@ -1,3 +1,0 @@
-vim.keymap.set('n', '<leader>u', vim.cmd.UndotreeToggle)
--- Allow for focus to change to undo tree on toggle
-vim.g.undotree_SetFocusWhenToggle = 1
