+++ .config/nvim/lua/senyc/home.lua
