+++ .config/nvim/lua/senyc/init.lua
@@ -1,4 +1,4 @@
 require('senyc.set') -- Needs to be called first to set global leader
-require('senyc.packer')
-require('senyc.visual')
+-- require('senyc.visual')
 require('senyc.remap')
+require('senyc.lazy')
