+++ .config/nvim/lua/senyc/plugins/alpha.lua
@@ -1,0 +1,8 @@
+return {
+
+  'goolord/alpha-nvim',
+  lazy = false,
+  config = function()
+    require('alpha').setup(require('senyc.home'))
+  end
+}
