+++ .config/nvim/lua/senyc/plugins/comment.lua
@@ -1,0 +1,8 @@
+return {
+  'numToStr/Comment.nvim',
+  keys = "gc",
+  lazy = true,
+  config = function()
+    require "Comment".setup()
+  end
+}
