+++ .config/nvim/lua/senyc/plugins/mason.lua
@@ -1,0 +1,11 @@
+return {
+  'williamboman/mason.nvim',
+  lazy = false,
+  config = function()
+    require 'mason'.setup {
+      ui = {
+        border = 'rounded'
+      }
+    }
+  end,
+}
