+++ .config/nvim/lua/senyc/plugins/nvim-conflict.lua
@@ -1,0 +1,27 @@
+return {
+  'akinsho/git-conflict.nvim',
+  tag = 'v1.0.0',
+
+  config = function()
+    require 'git-conflict'.setup {
+      default_mappings = true, -- disable buffer local mapping created by this plugin
+      default_commands = true, -- disable commands created by this plugin
+      disable_diagnostics = false, -- This will disable the diagnostics in a buffer whilst it is conflicted
+      list_opener = 'copen',   -- command or function to open the conflicts list
+      highlights = {           -- They must have background color, otherwise the default color will be used
+        incoming = 'DiffAdd',
+        current = 'DiffText',
+      }
+    }
+    vim.api.nvim_create_autocmd('User', {
+      pattern = 'GitConflictDetected',
+      callback = function()
+        vim.notify('Conflict detected in ' .. vim.fn.expand('<afile>'))
+        vim.keymap.set('n', 'cww', function()
+          engage.conflict_buster()
+          create_buffer_local_mappings()
+        end)
+      end
+    })
+  end
+}
