+++ .config/nvim/lua/senyc/plugins/treesitter-playground.lua
@@ -1,0 +1,6 @@
+return {
+  'nvim-treesitter/playground',
+  dependencies = { 'nvim-treesitter/nvim-treesitter' },
+  lazy = true,
+  cmd = 'TSPlaygroundToggle',
+}
