+++ .config/nvim/spell/en.utf-8.add
@@ -76,3 +76,5 @@ decrement
 formatoptions
 QuickFix
 macchiato
+first
+prefill
